Add log_exception() method.
log_exception assumes it is being called from an exception handler
context and creates an event from the result of sys.exc_info(), it also
takes an optional message argument to override whatever message is
contained within the exception.

log_error is used thoughout our codebase with just the exception
argument set, which is the case for which log_exception is intended.
There are a few places where log_error is not used in this way, with an
error class and message explicitly set, and where it's not called from
an exception handler context. Since backtrace defaults to true, it's
also reporting the backtrace of the last exception that occurred as the
backtrace for the event, which is incorrect.

log_error should be either depricated in favor of an explict call to
log, or the backtrace behavior should be removed.

Signed-off-by: Spiridon Eliopoulos <seliopou@gmail.com>