[settings] re-sync settings on each trace start (#108)
When oboe.config tracing_mode and sample_rate settings are updated, they are
pushed down into the oboe settings datastructures corresponding.  However, the
oboe settings are thread-local, while oboe.config is global.  This means that
if local configuration is done in a python thread other than the one that is
starting traces, the local settings will appear correct from the python
perspective but will not apply to tracing decisions.

This is a proposal to fix that by syncing the settings on each trace.