Do not fail if endEvt is None
This fixes a behaviour I've seen in this module when integrating with TurboGears and CherryPy.
An exception is raise because endEvt is false. This happens at the first request received by the application (but only at the first request).
I suspect the context starts as invalid, thus endEvt is not defined, but then becomes valid during runapp().

This proposed change fixes this behaviour. It might not produce a valid trace, but will avoid the library crashing.